Deploy ready

+++ src/main/java/br/unoeste/ativooperante/services/OrgaoService.java
@@ -1,9 +1,14 @@
 package br.unoeste.ativooperante.services;
 
+import br.unoeste.ativooperante.db.documents.Denuncia;
 import br.unoeste.ativooperante.db.documents.Orgao;
+import br.unoeste.ativooperante.db.documents.Tipo;
+import br.unoeste.ativooperante.db.repository.DenunciaRepository;
 import br.unoeste.ativooperante.db.repository.OrgaoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,6 +20,8 @@ public class OrgaoService {
     private OrgaoRepository orgaoRepository;
     @Autowired
     private MongoTemplate mongoTemplate;
+    @Autowired
+    private DenunciaRepository denunciaRepository;
 
     public Orgao save(Orgao orgao)
     {
@@ -31,15 +38,20 @@ public class OrgaoService {
         return this.orgaoRepository.findAll();
     }
 
-    public boolean delete (String id)
+    public ResponseEntity<Object> delete (String id)
     {
+        Orgao orgao = this.orgaoRepository.findById(id).orElse(null);
+        List<Denuncia> denuncias = this.denunciaRepository.findByOrgao(orgao);
         try {
+            if(!denuncias.isEmpty()) {
+                return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Há denúncias registradas com este órgão.");
+            }
             this.orgaoRepository.deleteById(id);
         }catch (Exception e)
         {
-            return false;
+            return ResponseEntity.badRequest().body("Erro ao excluir.");
         }
-        return true;
+        return ResponseEntity.ok("");
     }
 
     public Orgao alterar (String id,Orgao orgao)
